+++ src/utils/getUniqueShortcuts.ts
@@ -1,4 +1,4 @@
-import type { Shortcut } from '../models/shortcut';
+import type { Shortcut } from '@models/shortcut.js';
 
 export const getUniqueShortcuts = (shortcuts: Shortcut[]): Shortcut[] => {
   const items = new Set(shortcuts.map(shortcut => JSON.stringify(shortcut, null, 0)));
