+++ src/utils/hasModifier.ts
@@ -1,4 +1,4 @@
-import type { Shortcut } from '../models';
+import type { Shortcut } from '@models/shortcut.js';
 
 export const hasModifier = (event?: KeyboardEvent | Shortcut): boolean => {
   return event !== undefined && (!!event.altKey || !!event.ctrlKey || !!event.metaKey || !!event.shiftKey);
