+++ src/utils/type-predicate.ts
@@ -1,7 +1,7 @@
 import { isBoolean, isObject, isOptionalBoolean, isOptionalString, isString } from '@webdeveric/utils/type-predicate';
 import { everyItem } from '@webdeveric/utils/type-predicate-factory';
 
-import type { MetaKeys, Shortcut } from '../models/shortcut';
+import type { MetaKeys, Shortcut } from '@models/shortcut.js';
 
 const metaKeys: (keyof MetaKeys)[] = ['altKey', 'ctrlKey', 'metaKey', 'shiftKey'];
